moved widget.py to package widget
