+++ tox.ini
@@ -7,6 +7,7 @@ deps=
     pytest<=2.9.0,>=2.7.3
     pytest-localserver
     pytest-cov
+    anytree
 commands =
     py.test {posargs: -v --cov widgetastic --cov-report term-missing}
